Move OpenCodex dashboard source tests out of the oversized controller suite.
